Support `floattype(Rational)` and require <:AbstractFloat for fallback
It seems to be a non-sequitur to allow `floattype` to return a type that
is not an `AbstractFloat`, unless it has been extended as such.
The docs have been enhanced to clarify when it is OK to extend `floattype`
in ways that don't return an `AbstractFloat`.

Breaking change: no, not counting the deprecation warning.